refactor: integrate ImageCropperDialog into dialog system
- Add IMAGE_CROPPER to Dialog enum
- Move ImageCropperDialog to dialogs directory
- Refactor to use Redux dialog state management
- Update getCroppedImage to use original filename
- Update avatar/banner/logo managers to use dialog system

Addresses PR #553 review feedback
