+++ apps/client/src/components/dialogs/dialogs.ts
@@ -10,5 +10,6 @@ export enum Dialog {
   PLUGIN_COMMANDS = 'PLUGIN_COMMANDS',
   PLUGIN_SETTINGS = 'PLUGIN_SETTINGS',
   DELETE_USER = 'DELETE_USER',
-  SEARCH = 'SEARCH'
+  SEARCH = 'SEARCH',
+  IMAGE_CROPPER = 'IMAGE_CROPPER'
 }
