+++ apps/client/src/components/server-screens/server-settings/general/logo-manager.tsx
@@ -1,11 +1,12 @@
-import { ImageCropperDialog } from '@/components/image-cropper-dialog';
+import { Dialog } from '@/components/dialogs/dialogs';
 import { ImagePicker } from '@/components/image-picker';
+import { openDialog } from '@/features/dialogs/actions';
 import { uploadFile } from '@/helpers/upload-file';
 import { useFilePicker } from '@/hooks/use-file-picker';
 import { getTRPCClient } from '@/lib/trpc';
 import type { TFile } from '@sharkord/shared';
 import { Group } from '@sharkord/ui';
-import { memo, useCallback, useState } from 'react';
+import { memo, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { toast } from 'sonner';
 
@@ -17,8 +18,6 @@ type TLogoManagerProps = {
 const LogoManager = memo(({ logo, refetch }: TLogoManagerProps) => {
   const { t } = useTranslation('dialogs');
   const openFilePicker = useFilePicker();
-  const [cropperOpen, setCropperOpen] = useState(false);
-  const [pendingImageSrc, setPendingImageSrc] = useState<string | null>(null);
 
   const removeLogo = useCallback(async () => {
     const trpc = getTRPCClient();
@@ -33,21 +32,6 @@ const LogoManager = memo(({ logo, refetch }: TLogoManagerProps) => {
       toast.error(t('logoRemoveFailed'));
     }
   }, [refetch, t]);
-
-  const onLogoClick = useCallback(async () => {
-    try {
-      const [file] = await openFilePicker('image/*');
-
-      const reader = new FileReader();
-      reader.onload = () => {
-        setPendingImageSrc(reader.result as string);
-        setCropperOpen(true);
-      };
-      reader.readAsDataURL(file);
-    } catch {
-      // user cancelled
-    }
-  }, [openFilePicker]);
 
   const onCropConfirm = useCallback(
     async (croppedFile: File) => {
@@ -67,42 +51,41 @@ const LogoManager = memo(({ logo, refetch }: TLogoManagerProps) => {
         toast.success(t('logoUpdatedSuccess'));
       } catch {
         toast.error(t('logoUpdateFailed'));
-      } finally {
-        setPendingImageSrc(null);
       }
     },
     [refetch, t]
   );
 
-  return (
-    <>
-      <Group
-        label={t('logoLabel')}
-        description={t('logoRecommendedResolution')}
-      >
-        <ImagePicker
-          image={logo}
-          onImageClick={onLogoClick}
-          onRemoveImageClick={removeLogo}
-          className="w-48 h-48"
-        />
-      </Group>
+  const onLogoClick = useCallback(async () => {
+    try {
+      const [file] = await openFilePicker('image/*');
 
-      {pendingImageSrc && (
-        <ImageCropperDialog
-          open={cropperOpen}
-          onClose={() => {
-            setCropperOpen(false);
-            setPendingImageSrc(null);
-          }}
-          imageSrc={pendingImageSrc}
-          aspect={1}
-          cropShape="rect"
-          title={t('cropLogoTitle')}
-          onConfirm={onCropConfirm}
-        />
-      )}
-    </>
+      const reader = new FileReader();
+      reader.onload = () => {
+        openDialog(Dialog.IMAGE_CROPPER, {
+          imageSrc: reader.result as string,
+          originalFileName: file.name,
+          aspect: 1,
+          cropShape: 'rect',
+          title: t('cropLogoTitle'),
+          onConfirm: onCropConfirm
+        });
+      };
+      reader.readAsDataURL(file);
+    } catch {
+      // user cancelled
+    }
+  }, [openFilePicker, onCropConfirm, t]);
+
+  return (
+    <Group label={t('logoLabel')} description={t('logoRecommendedResolution')}>
+      <ImagePicker
+        image={logo}
+        onImageClick={onLogoClick}
+        onRemoveImageClick={removeLogo}
+        className="w-48 h-48"
+      />
+    </Group>
   );
 });
